requirements.txt
# Core AI/ML libraries
openai>=1.0.0
anthropic>=0.3.0
google-genai

# PDF processing
PyMuPDF>=1.23.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0

# Video/Audio processing
moviepy>=1.0.3
pydub>=0.25.1

# Audio transcription (for batch splitting)
SpeechRecognition>=3.10.0

# Sorting and utilities
natsort>=8.0.0

# HTTP requests
requests>=2.28.0

# Optional: GUI development (uncomment if needed)
# tkinter (usually comes with Python)
# PyQt5>=5.15.0
# kivy>=2.1.0

# Optional: Web interface (uncomment if needed)  
# flask>=2.0.0
# fastapi>=0.68.0
# uvicorn>=0.15.0
python-dotenv
